+++ app/src/main/java/jp/co/tbdeveloper/warikanapp/feature_roulette/data/repository/RouletteRepositoryImpl.kt
@@ -12,7 +12,7 @@ class RouletteRepositoryImpl(
         return dao.getRoulettes()
     }
 
-    override suspend fun getRouletteById(id: Int): RouletteEntity? {
+    override suspend fun getRouletteById(id: Long): RouletteEntity? {
         return dao.getRouletteById(id)
     }
